requirements.txt
pydantic==2.8.2
starlette==0.37.2
uvicorn==0.30.3
fastapi==0.111.1
PyJWT==2.8.0
pandas==2.2.2
pyarrow==17.0.0
fastparquet==2024.5.0
passlib==1.7.4
pytest==8.3.2
pytest-mock==3.14.0
